Importing information from the tRNA db as GRanges object

title

TRNA_DB_URL

TRNA_DB_URL_MT

import.tRNAdb.id(
  tdbID,
  database = c("DNA", "RNA"),
  origin = c("allothers", "plastid", "mitochondrial"),
  dbURL = TRNA_DB_URL,
  verbose = FALSE
)

import.mttRNAdb.id(mtdbID, dbURL = TRNA_DB_URL_MT, verbose = FALSE)

import.tRNAdb.blast(
  blastSeq,
  database = c("DNA", "RNA"),
  origin = c("allothers", "plastid", "mitochondrial"),
  dbURL = TRNA_DB_URL,
  verbose = FALSE
)

import.tRNAdb(
  organism = "",
  strain = "",
  taxonomyID = "",
  aminoacids = "",
  anticodons = "",
  sequences = list(),
  structures = list(),
  reference = "",
  comment = "",
  pubmed = "",
  genes = "",
  database = c("DNA", "RNA"),
  origin = c("allothers", "plastid", "mitochondrial"),
  dbURL = TRNA_DB_URL,
  verbose = FALSE
)

import.mttRNAdb(
  organism = "",
  strain = "",
  taxonomyID = "",
  aminoacids = "",
  anticodons = "",
  sequences = list(),
  structures = list(),
  reference = "",
  comment = "",
  pubmed = "",
  genes = "",
  dbURL = TRNA_DB_URL_MT,
  verbose = FALSE
)

tRNAdb2GFF(input)

Format

An object of class character of length 1.

An object of class character of length 1.

Arguments

tdbID

a tRNAdb ID

database

"RNA" or "DNA"

origin

one ore more of "plastid", "mitochondrial" or "allothers"

dbURL

the URL of the tRNA db

verbose

whether to report verbose information from the httr2 calls

mtdbID

a mtRNAdb ID

blastSeq

a sequence to use for a blast search

organism

a organism name as a character string

strain

a strain information as a character string

taxonomyID

organism and strain information as a taxonom ID

aminoacids

a character vector of amino acids as a three letter code

anticodons

a character vector of anticodon sequences

sequences

a named (1-15) list of sequences, which are used for the search

structures

a named (1-15) list of structures, which are used for the search. Please use the \(\) or >< dot bracket annotation.

reference

a reference as a character string

comment

a comment as a character string

pubmed

a pubmed ID

genes

a gene name as a character string

input

a GRanges object which passes the istRNAdbGRanges check

Value

a GRanges object containing the information from the tRNA db

Examples

import.tRNAdb(organism = "Saccharomyces cerevisiae",
              aminoacids = c("Phe","Ala"))
#> Warning: tRNAdb Server seems to be not available.
#> GRanges object with 0 ranges and 0 metadata columns:
#>    seqnames    ranges strand
#>       <Rle> <IRanges>  <Rle>
#>   -------
#>   seqinfo: no sequences
import.tRNAdb.id(tdbID = "tdbD00000785")
#> Warning: tRNAdb Server seems to be not available.
#> GRanges object with 0 ranges and 0 metadata columns:
#>    seqnames    ranges strand
#>       <Rle> <IRanges>  <Rle>
#>   -------
#>   seqinfo: no sequences
import.tRNAdb.blast(blastSeq =
"GCGGATTTAGCTCAGTTGGGAGAGCGCCAGACTGAAGATCTGGAGGTCCTGTGTTCGATCCACAGAATTCGCA")
#> Warning: tRNAdb Server seems to be not available.
#> GRanges object with 0 ranges and 0 metadata columns:
#>    seqnames    ranges strand
#>       <Rle> <IRanges>  <Rle>
#>   -------
#>   seqinfo: no sequences
import.mttRNAdb(organism = "Bos taurus",
                aminoacids = c("Phe","Ala"))
#> Warning: tRNAdb Server seems to be not available.
#> GRanges object with 0 ranges and 0 metadata columns:
#>    seqnames    ranges strand
#>       <Rle> <IRanges>  <Rle>
#>   -------
#>   seqinfo: no sequences
import.mttRNAdb.id(mtdbID = "mtdbD00000900")
#> Warning: tRNAdb Server seems to be not available.
#> GRanges object with 0 ranges and 0 metadata columns:
#>    seqnames    ranges strand
#>       <Rle> <IRanges>  <Rle>
#>   -------
#>   seqinfo: no sequences