Open a tRNA db entry in a browser

open.tdbID is a wrapper for browseURL and opens a tab for a tRNAdb entry in a browser. Please note, that the tRNAdb server does not show the entry right away without a session ID. open twice upon first use.

open_tdbID(tdbID, dbURL = TRNA_DB_URL)

open_mtdbID(mtdbID, dbURL = TRNA_DB_URL_MT)

Arguments

tdbID

a tRNA db

dbURL

the URL for the tRNAdb

mtdbID

a mtRNA db

Value

opens a window in a default browser for tRNAdb entry selected

Examples

if(interactive()){
  open_tdbID("tdbD00000785")
  open_mtdbID("mtdbD00000907")
}