tRNAdbImport: Importing from to tRNAdb and mitotRNAdb as GRanges

The tRNAdb and mttRNAdb (Jühling et al. 2009) is a compilation of tRNA sequences and tRNA genes. It is a follow up version of the database of Sprinzl et al. 2005.

Using `tRNAdbImport` the tRNAdb can be accessed as outlined on the website [http://trna.bioinf.uni-leipzig.de/](http://trna.bioinf.uni-leipzig.de/) and the results are returned as a `GRanges` object.

Manual

Please refer to the tRNAdbImport vignette for an example how to work and use the package: tRNAdbImport

References

Jühling F, Mörl M, Hartmann RK, Sprinzl M, Stadler PF, Pütz J. 2009. "tRNAdb 2009: compilation of tRNA sequences and tRNA genes." Nucleic Acids Research, Volume 37 (suppl_1): D159–162. doi:10.1093/nar/gkn772.

See also

Author

Felix G M Ernst [aut]