Importing information from the tRNA db as GRanges object

title

TRNA_DB_URL

TRNA_DB_URL_MT

import.tRNAdb.id(
  tdbID,
  database = c("DNA", "RNA"),
  origin = c("allothers", "plastid", "mitochondrial"),
  dbURL = TRNA_DB_URL,
  verbose = FALSE
)

import.mttRNAdb.id(mtdbID, dbURL = TRNA_DB_URL_MT, verbose = FALSE)

import.tRNAdb.blast(
  blastSeq,
  database = c("DNA", "RNA"),
  origin = c("allothers", "plastid", "mitochondrial"),
  dbURL = TRNA_DB_URL,
  verbose = FALSE
)

import.tRNAdb(
  organism = "",
  strain = "",
  taxonomyID = "",
  aminoacids = "",
  anticodons = "",
  sequences = list(),
  structures = list(),
  reference = "",
  comment = "",
  pubmed = "",
  genes = "",
  database = c("DNA", "RNA"),
  origin = c("allothers", "plastid", "mitochondrial"),
  dbURL = TRNA_DB_URL,
  verbose = FALSE
)

import.mttRNAdb(
  organism = "",
  strain = "",
  taxonomyID = "",
  aminoacids = "",
  anticodons = "",
  sequences = list(),
  structures = list(),
  reference = "",
  comment = "",
  pubmed = "",
  genes = "",
  dbURL = TRNA_DB_URL_MT,
  verbose = FALSE
)

tRNAdb2GFF(input)

Arguments

tdbID
a tRNAdb ID

database
"RNA" or "DNA"

origin
one ore more of "plastid", "mitochondrial" or "allothers"

dbURL
the URL of the tRNA db

verbose
whether to report verbose information from the httr calls

mtdbID
a mtRNAdb ID

blastSeq
a sequence to use for a blast search

organism
a organism name as a character string

strain
a strain information as a character string

taxonomyID
organism and strain information as a taxonom ID

aminoacids
a character vector of amino acids as a three letter code

anticodons
a character vector of anticodon sequences

sequences
a named (1-15) list of sequences, which are used for the search

structures
a named (1-15) list of structures, which are used for the search. Please use the \(\) or >< dot bracket annotation.

reference
a reference as a character string

comment
a comment as a character string

pubmed
a pubmed ID

genes
a gene name as a character string

input
a GRanges object which passes the istRNAdbGRanges check

Format

An object of class character of length 1.

Value

a GRanges object containing the information from the tRNA db

Examples

import.tRNAdb(organism = "Saccharomyces cerevisiae",
              aminoacids = c("Phe","Ala"))

#> GRanges object with 13 ranges and 15 metadata columns:
#>            seqnames    ranges strand |        no tRNA_length   tRNA_type
#>               <Rle> <IRanges>  <Rle> | <integer>   <integer> <character>
#>    [1] tdbD00000218      1-73      * |         1          73         Ala
#>    [2] tdbD00000219      1-73      * |         2          73         Ala
#>    [3] tdbD00000785      1-73      * |         3          73         Phe
#>    [4] tdbD00000786      1-73      * |         4          73         Phe
#>    [5] tdbD00000787      1-73      * |         5          73         Phe
#>    ...          ...       ...    ... .       ...         ...         ...
#>    [9] tdbD00005005      1-73      * |         9          73         Phe
#>   [10] tdbD00005006      1-73      * |        10          73         Phe
#>   [11] tdbD00005007      1-73      * |        11          73         Phe
#>   [12] tdbD00005008      1-73      * |        12          73         Phe
#>   [13] tdbD00005009      1-73      * |        13          73         Phe
#>        tRNA_anticodon                tRNA_seq                tRNA_str
#>           <character>          <DNAStringSet>   <DotBracketStringSet>
#>    [1]            TGC GGGCACATGG...GTTGCGTCCA <<<<.<<..<...>>>>.>>>>.
#>    [2]            AGC GGGCGTGTGG...GACTCGTCCA <<<<<.<..<...>>>.>>>>>.
#>    [3]            GAA GCGGATTTAG...AGAATTCGCA <<<<<<<..<...>>>>>>>>>.
#>    [4]            GAA GCGGATTTAG...AGAGTTCGCA <<<<<<<..<...>>>>>>>>>.
#>    [5]            GAA GCGGATTTAG...AGAATTCGCA <<<<<<<..<...>>>>>>>>>.
#>    ...            ...                     ...                     ...
#>    [9]            GAA GCGGACTTAG...AGAGTTCGCA <<<<<<<..<...>>>>>>>>>.
#>   [10]            GAA GCGGATTTAG...AGAATTCGCA <<<<<<<..<...>>>>>>>>>.
#>   [11]            GAA GCGGATTTAG...AGAATTCGCA <<<<<<<..<...>>>>>>>>>.
#>   [12]            GAA GCGGATTTAG...AGAATTCGCA <<<<<<<..<...>>>>>>>>>.
#>   [13]            GAA GCGGACTTAG...AGAGTTCGCA <<<<<<<..<...>>>>>>>>>.
#>        tRNA_CCA.end      tRNAdb    tRNAdb_ID          tRNAdb_organism
#>           <logical> <character>  <character>              <character>
#>    [1]        FALSE         DNA tdbD00000218 Saccharomyces cerevisiae
#>    [2]        FALSE         DNA tdbD00000219 Saccharomyces cerevisiae
#>    [3]        FALSE         DNA tdbD00000785 Saccharomyces cerevisiae
#>    [4]        FALSE         DNA tdbD00000786 Saccharomyces cerevisiae
#>    [5]        FALSE         DNA tdbD00000787 Saccharomyces cerevisiae
#>    ...          ...         ...          ...                      ...
#>    [9]        FALSE         DNA tdbD00005005 Saccharomyces cerevisiae
#>   [10]        FALSE         DNA tdbD00005006 Saccharomyces cerevisiae
#>   [11]        FALSE         DNA tdbD00005007 Saccharomyces cerevisiae
#>   [12]        FALSE         DNA tdbD00005008 Saccharomyces cerevisiae
#>   [13]        FALSE         DNA tdbD00005009 Saccharomyces cerevisiae
#>        tRNAdb_strain tRNAdb_taxonomyID tRNAdb_verified
#>          <character>       <character>       <logical>
#>    [1]                            4932           FALSE
#>    [2]                            4932           FALSE
#>    [3]                            4932           FALSE
#>    [4]                            4932           FALSE
#>    [5]                            4932           FALSE
#>    ...           ...               ...             ...
#>    [9]           unk              4932           FALSE
#>   [10]           unk              4932           FALSE
#>   [11]           unk              4932           FALSE
#>   [12]           unk              4932           FALSE
#>   [13]           unk              4932           FALSE
#>                                                                                                                                              tRNAdb_reference
#>                                                                                                                                               <CharacterList>
#>    [1]                                                                                      H.J.DRABKIN, U.L.RAJBHANDARY (1985) J. BIOL. CHEM. 260, 5596-5602
#>    [2]                                                                 F.CREUSOT, M.GAISNE, J.VERDIERE, P.P.SLONIMSKI (1989) NUCLEIC ACIDS RES. 17, 1865-1866
#>    [3]          P.VALENZUELA, A.VENEGAS, F.WEINBERG, R.BISHOP, W.J.RUTTER (1978) PROC. NATL. ACAD. SCI. USA 75, 190-194;  P.BULL ET AL. (1987) DNA 6, 353-362
#>    [4]                                                                                                                    P.BULL ET AL. (1987) DNA 6, 353-362
#>    [5]                                                                                                           A.GOFFEAU ET AL. (1996) SCIENCE 274, 546-567
#>    ...                                                                                                                                                    ...
#>    [9] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>   [10] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>   [11] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>   [12] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>   [13] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>            tRNAdb_pmid
#>        <CharacterList>
#>    [1]                
#>    [2]                
#>    [3]                
#>    [4]                
#>    [5]                
#>    ...             ...
#>    [9]         9023104
#>   [10]         9023104
#>   [11]         9023104
#>   [12]         9023104
#>   [13]         9023104
#>   -------
#>   seqinfo: 13 sequences from an unspecified genome; no seqlengths

import.tRNAdb.id(tdbID = "tdbD00000785")

#> GRanges object with 1 range and 15 metadata columns:
#>           seqnames    ranges strand |        no tRNA_length   tRNA_type
#>              <Rle> <IRanges>  <Rle> | <integer>   <integer> <character>
#>   [1] tdbD00000785      1-73      * |         1          73         Phe
#>       tRNA_anticodon                tRNA_seq                tRNA_str
#>          <character>          <DNAStringSet>   <DotBracketStringSet>
#>   [1]            GAA GCGGATTTAG...AGAATTCGCA <<<<<<<..<...>>>>>>>>>.
#>       tRNA_CCA.end      tRNAdb    tRNAdb_ID          tRNAdb_organism
#>          <logical> <character>  <character>              <character>
#>   [1]        FALSE         DNA tdbD00000785 Saccharomyces cerevisiae
#>       tRNAdb_strain tRNAdb_taxonomyID tRNAdb_verified
#>         <character>       <character>       <logical>
#>   [1]                            4932           FALSE
#>                                                                                                                                    tRNAdb_reference
#>                                                                                                                                     <CharacterList>
#>   [1] P.VALENZUELA, A.VENEGAS, F.WEINBERG, R.BISHOP, W.J.RUTTER (1978) PROC. NATL. ACAD. SCI. USA 75, 190-194;  P.BULL ET AL. (1987) DNA 6, 353-362
#>           tRNAdb_pmid
#>       <CharacterList>
#>   [1]                
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

import.tRNAdb.blast(blastSeq =
"GCGGATTTAGCTCAGTTGGGAGAGCGCCAGACTGAAGATCTGGAGGTCCTGTGTTCGATCCACAGAATTCGCA")

#> GRanges object with 10 ranges and 15 metadata columns:
#>            seqnames    ranges strand |        no tRNA_length   tRNA_type
#>               <Rle> <IRanges>  <Rle> | <integer>   <integer> <character>
#>    [1] tdbD00000785      1-73      * |         1          73         Phe
#>    [2] tdbD00000786      1-73      * |         2          73         Phe
#>    [3] tdbD00005005      1-73      * |         3          73         Phe
#>    [4] tdbD00005006      1-73      * |         4          73         Phe
#>    [5] tdbD00005007      1-73      * |         5          73         Phe
#>    [6] tdbD00005008      1-73      * |         6          73         Phe
#>    [7] tdbD00005009      1-73      * |         7          73         Phe
#>    [8] tdbD00005094      1-73      * |         8          73         Phe
#>    [9] tdbR00000083      1-76      * |         9          76         Phe
#>   [10] tdbR00000084      1-76      * |        10          76         Phe
#>        tRNA_anticodon                tRNA_seq                tRNA_str
#>           <character>       <ModRNAStringSet>   <DotBracketStringSet>
#>    [1]            GAA GCGGAUUUAG...AGAAUUCGCA <<<<<<<..<...>>>>>>>>>.
#>    [2]            GAA GCGGAUUUAG...AGAGUUCGCA <<<<<<<..<...>>>>>>>>>.
#>    [3]            GAA GCGGACUUAG...AGAGUUCGCA <<<<<<<..<...>>>>>>>>>.
#>    [4]            GAA GCGGAUUUAG...AGAAUUCGCA <<<<<<<..<...>>>>>>>>>.
#>    [5]            GAA GCGGAUUUAG...AGAAUUCGCA <<<<<<<..<...>>>>>>>>>.
#>    [6]            GAA GCGGAUUUAG...AGAAUUCGCA <<<<<<<..<...>>>>>>>>>.
#>    [7]            GAA GCGGACUUAG...AGAGUUCGCA <<<<<<<..<...>>>>>>>>>.
#>    [8]            GAA GCGGUUUUAG...AGAAAUCGCA <<<<<<<..<...>>>>>>>>>.
#>    [9]            #AA GCGGAUUUAL...AUUCGCACCA <<<<<<<..<...>>>>>>....
#>   [10]            #AA GCGGACUUAL...GUUCGCACCA <<<<<<<..<...>>>>>>....
#>        tRNA_CCA.end      tRNAdb    tRNAdb_ID          tRNAdb_organism
#>           <logical> <character>  <character>              <character>
#>    [1]        FALSE         DNA tdbD00000785 Saccharomyces cerevisiae
#>    [2]        FALSE         DNA tdbD00000786 Saccharomyces cerevisiae
#>    [3]        FALSE         DNA tdbD00005005 Saccharomyces cerevisiae
#>    [4]        FALSE         DNA tdbD00005006 Saccharomyces cerevisiae
#>    [5]        FALSE         DNA tdbD00005007 Saccharomyces cerevisiae
#>    [6]        FALSE         DNA tdbD00005008 Saccharomyces cerevisiae
#>    [7]        FALSE         DNA tdbD00005009         Candida glabrata
#>    [8]        FALSE         DNA tdbD00005094 Saccharomyces cerevisiae
#>    [9]         TRUE         RNA tdbR00000083 Saccharomyces cerevisiae
#>   [10]         TRUE         RNA tdbR00000084 Saccharomyces cerevisiae
#>          tRNAdb_strain tRNAdb_taxonomyID tRNAdb_verified
#>            <character>       <character>       <logical>
#>    [1]                              4932           FALSE
#>    [2]             unk              4932           FALSE
#>    [3]             unk              4932           FALSE
#>    [4]             unk              4932           FALSE
#>    [5]             unk              4932           FALSE
#>    [6]             unk              4932           FALSE
#>    [7] CBS 138, CBS138              4932           FALSE
#>    [8]                            284593           FALSE
#>    [9]                              4932            TRUE
#>   [10]                              4932            TRUE
#>                                                                                                                                              tRNAdb_reference
#>                                                                                                                                               <CharacterList>
#>    [1]          P.VALENZUELA, A.VENEGAS, F.WEINBERG, R.BISHOP, W.J.RUTTER (1978) PROC. NATL. ACAD. SCI. USA 75, 190-194;  P.BULL ET AL. (1987) DNA 6, 353-362
#>    [2]                                                                                                                    P.BULL ET AL. (1987) DNA 6, 353-362
#>    [3] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>    [4] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>    [5] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>    [6] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>    [7] Lowe, T.M. & Eddy, S.R. (1997) "tRNAscan-SE: A program for improved detection of transfer RNA genes in genomic sequence" Nucl. Acids Res. 25: 955-964.
#>    [8]                                                                                 Dujon B. et al. (2004) "Genome evolution in yeasts", Nature 430:35-44.
#>    [9]                                                                                  P.E.NIELSEN, V.LEICK (1982) BIOCHEM. BIOPHYS. RES. COMM. 106, 179-185
#>   [10]                                                                                  G.KEITH, G.DIRHEIMER (1987) BIOCHEM. BIOPHYS. RES. COMM. 142, 183-187
#>            tRNAdb_pmid
#>        <CharacterList>
#>    [1]                
#>    [2]                
#>    [3]         9023104
#>    [4]         9023104
#>    [5]         9023104
#>    [6]         9023104
#>    [7]         9023104
#>    [8]               0
#>    [9]                
#>   [10]                
#>   -------
#>   seqinfo: 10 sequences from an unspecified genome; no seqlengths

import.mttRNAdb(organism = "Bos taurus",
                aminoacids = c("Phe","Ala"))

#> GRanges object with 2 ranges and 15 metadata columns:
#>            seqnames    ranges strand |        no tRNA_length   tRNA_type
#>               <Rle> <IRanges>  <Rle> | <integer>   <integer> <character>
#>   [1] mtdbD00000900      1-67      * |         1          67         Phe
#>   [2] mtdbD00000907      1-69      * |         2          69         Ala
#>       tRNA_anticodon                tRNA_seq                tRNA_str
#>          <character>          <DNAStringSet>   <DotBracketStringSet>
#>   [1]            GAA GTTGATGTAG...TCCATAAACA <<<.<<<..<...>>>>>.>>>.
#>   [2]            TGC GAGGATTTAG...GCAATCCTTA <<<<<<<..<...>>>>>>>>>.
#>       tRNA_CCA.end      tRNAdb     tRNAdb_ID tRNAdb_organism tRNAdb_strain
#>          <logical> <character>   <character>     <character>   <character>
#>   [1]        FALSE          MT mtdbD00000900      Bos taurus              
#>   [2]        FALSE          MT mtdbD00000907      Bos taurus              
#>       tRNAdb_taxonomyID tRNAdb_verified   tRNAdb_reference     tRNAdb_pmid
#>             <character>       <logical>    <CharacterList> <CharacterList>
#>   [1]              9913           FALSE NC_006853.1 (NCBI)                
#>   [2]              9913           FALSE NC_006853.1 (NCBI)                
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

import.mttRNAdb.id(mtdbID = "mtdbD00000900")

#> GRanges object with 1 range and 15 metadata columns:
#>            seqnames    ranges strand |        no tRNA_length   tRNA_type
#>               <Rle> <IRanges>  <Rle> | <integer>   <integer> <character>
#>   [1] mtdbD00000900      1-67      * |         1          67         Phe
#>       tRNA_anticodon                tRNA_seq                tRNA_str
#>          <character>          <DNAStringSet>   <DotBracketStringSet>
#>   [1]            GAA GTTGATGTAG...TCCATAAACA <<<.<<<..<...>>>>>.>>>.
#>       tRNA_CCA.end      tRNAdb     tRNAdb_ID tRNAdb_organism tRNAdb_strain
#>          <logical> <character>   <character>     <character>   <character>
#>   [1]        FALSE          MT mtdbD00000900      Bos taurus              
#>       tRNAdb_taxonomyID tRNAdb_verified   tRNAdb_reference     tRNAdb_pmid
#>             <character>       <logical>    <CharacterList> <CharacterList>
#>   [1]              9913           FALSE NC_006853.1 (NCBI)                
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths